end of section 10: toastify, errrorhandling

+++ Application/Activities/Delete.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Threading;
 using System.Threading.Tasks;
+using Application.Core;
 using Domain;
 using MediatR;
 using Persistence;
@@ -9,12 +10,12 @@ namespace Application.Activities
 {
     public class Delete
     {
-        public class Command : IRequest
+        public class Command : IRequest<Result<Unit>>
         {
             public Guid Id { get; set; }
         }
 
-        public class Handler : IRequestHandler<Command>
+        public class Handler : IRequestHandler<Command, Result<Unit>>
         {
             private readonly DataContext context;
 
@@ -23,16 +24,20 @@ namespace Application.Activities
                 this.context = context;
             }
 
-            public async Task<Unit> Handle(Command request, CancellationToken cancellationToken)
+            public async Task<Result<Unit>> Handle(Command request, CancellationToken cancellationToken)
             {
                 Activity activity = await context.Activities.FindAsync(request.Id);
 
-                //TODO: null check activity
+                if (activity == null)
+                    return null;
+
                 context.Activities.Remove(activity);
+                var deleted = await context.SaveChangesAsync() > 0;
 
-                await context.SaveChangesAsync();
+                if (deleted)
+                    return Result<Unit>.Success(Unit.Value);
 
-                return Unit.Value;
+                return Result<Unit>.Failure("Failed to delete activity");
             }
         }
     }
